Multiple Updates / Bug fixes
a) Ensure MQTT 3,4 & 5 correctly close the transactional context when committing/rolling back
b) Stop the mDNS registery in parallel to the server shutdown. It takes time to deregister entries
c) On startup the destination subscription state is validated with the resource. This will clean up any mismatch in stored event or subscribed event
d) Stop the UDP advertising (mqtt-sn) from binding to 0.0.0.0 ( or any-address )
e) Update the library versions